support/pyproject.toml
[tool.poetry]
name = "support"
version = "0.1.0"
description = ""
authors = ["razmanov666 <razmanov666@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
djoser = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
